refactor: move filles to new filled_tree directory
