+++ crates/committer/src/patricia_merkle_tree.rs
@@ -1,10 +1,8 @@
 pub mod errors;
-pub mod filled_node;
 pub mod filled_tree;
 pub mod original_skeleton_calc;
 pub mod original_skeleton_node;
 pub mod original_skeleton_tree;
-pub mod serialized_node;
 pub mod types;
 pub mod updated_skeleton_node;
 pub mod updated_skeleton_tree;
