+++ crates/committer/src/patricia_merkle_tree/types.rs
@@ -1,7 +1,7 @@
 use std::marker::PhantomData;
 
 use crate::hash::hash_trait::{HashFunction, HashInputPair, HashOutput};
-use crate::patricia_merkle_tree::filled_node::{BinaryData, LeafData, NodeData};
+use crate::patricia_merkle_tree::filled_tree::node::{BinaryData, LeafData, NodeData};
 use crate::types::Felt;
 
 #[cfg(test)]
