+++ crates/committer/src/patricia_merkle_tree/types_test.rs
@@ -1,6 +1,6 @@
 use crate::hash::hash_trait::{HashFunction, HashInputPair, HashOutput};
 use crate::hash::pedersen::PedersenHashFunction;
-use crate::patricia_merkle_tree::filled_node::{BinaryData, NodeData};
+use crate::patricia_merkle_tree::filled_tree::node::{BinaryData, NodeData};
 use crate::patricia_merkle_tree::types::EdgeData;
 use crate::patricia_merkle_tree::types::TreeHashFunction;
 use crate::patricia_merkle_tree::types::{
